Rename pyagg -> celiagg [ci skip]
